@@ -1,10 +1,9 @@
 package com.ciphertext.medicalinformationbackend.model;
 
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.*;
-import java.util.Date;
 
 /**
  * @author Sadman
